refactor : MVC에 맞게 디렉터리 변경
- model은 domain패키지 아래에 존재
- controller는 Main만을 가지고 있다
- view는 view패키지 아래에 존재
