+++ src/main/java/domain/Car.java
@@ -1,4 +1,4 @@
-package racing;
+package domain;
 
 public class Car {
